introduce OpenAI Mixin class for VectorIO
